requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.25.1
pandas==2.1.1
googlemaps==4.10.0
fpdf==1.7.2
beautifulsoup4==4.12.2
aiofiles==23.2.1
indic-nlp-library==0.91
python-multipart==0.0.6
pymongo==4.5.0
requests==2.31.0
